fix(cli): output short ID instead of full ID in quiet mode
The -q flag on all commands was printing the full 32-character ID.
Now outputs the 8-character short ID for consistency with how IDs
are displayed everywhere else in the CLI.

+++ src/cli/commands/artifact/create.rs
@@ -112,7 +112,7 @@ impl Command {
     }
 
     if self.quiet {
-      println!("{}", artifact.id);
+      println!("{}", artifact.id.short());
       return Ok(());
     }
 
@@ -167,7 +167,7 @@ impl Command {
       }
 
       if self.quiet {
-        println!("{}", artifact.id);
+        println!("{}", artifact.id.short());
       } else {
         let json = serde_json::to_string(&artifact)?;
         println!("{json}");
