+++ src/cli/commands/iteration/create.rs
@@ -66,7 +66,7 @@ impl Command {
     }
 
     if self.quiet {
-      println!("{}", iteration.id);
+      println!("{}", iteration.id.short());
       return Ok(());
     }
