+++ src/cli/commands/iteration/update.rs
@@ -75,7 +75,7 @@ impl Command {
     }
 
     if self.quiet {
-      println!("{}", iteration.id);
+      println!("{}", iteration.id.short());
       return Ok(());
     }
