+++ src/cli/commands/task/meta/set.rs
@@ -32,7 +32,7 @@ impl Command {
     if self.json {
       println!("{}", serde_json::to_string_pretty(&task)?);
     } else if self.quiet {
-      println!("{}", task.id);
+      println!("{}", task.id.short());
     } else {
       let msg = format!("Set {}.{} = {}", task.id, self.path, self.value);
       println!("{}", SuccessMessage::new(&msg, &ctx.theme));
