+++ src/cli/commands/task/update.rs
@@ -90,7 +90,7 @@ impl Command {
     }
 
     if self.quiet {
-      println!("{}", task.id);
+      println!("{}", task.id.short());
       return Ok(());
     }
